# Conflicts:
#   Backend/src/main/java/com/baekopa/backend/domain/meeting/controller/MeetingResultController.java
#   Backend/src/main/java/com/baekopa/backend/domain/meeting/repository/RemindQuizRepository.java
#   Backend/src/main/java/com/baekopa/backend/domain/meeting/service/MeetingService.java

+++ Backend/src/main/java/com/baekopa/backend/domain/meeting/service/MeetingService.java
@@ -14,6 +14,7 @@ import com.baekopa.backend.domain.meeting.repository.MeetingRepository;
 import com.baekopa.backend.domain.meeting.repository.MeetingScriptRepository;
 import com.baekopa.backend.domain.meeting.repository.MeetingSummaryRepository;
 import com.baekopa.backend.domain.member.entity.Member;
+import com.baekopa.backend.domain.note.dto.SubmittedNoteDto;
 import com.baekopa.backend.domain.note.entity.Note;
 import com.baekopa.backend.domain.note.entity.SubmittedNote;
 import com.baekopa.backend.domain.note.repository.SubmittedNoteRepository;
@@ -38,6 +39,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
+import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -97,7 +100,7 @@ public class MeetingService {
 
     public MeetingResponseDTO getMeetingResultAll(Long meetingId) {
         // 미팅 정보
-        Meeting meeting = meetingRepository.findById(meetingId)
+        Meeting meeting = meetingRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         // 미팅 전문
         MeetingScript meetingScript = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meeting)
@@ -113,6 +116,13 @@ public class MeetingService {
             meetingKeywordDTOList.add(meetingKeywordDTO);
         }
 
+        //제출된 개인 요약, 전체 요약 조회
+        List<SubmittedNoteDto> submittedNoteDtoList = submittedNoteRepository.findAllByMeetingAndDeletedAtIsNull(meeting)
+                .stream().map(SubmittedNoteDto::of).toList();
+
+        MeetingSubmittedNoteResponseDto submittedNoteSummary = MeetingSubmittedNoteResponseDto.of(submittedNoteDtoList, meeting.getNoteSummary());
+
+
         // 미팅 참여자
         List<SubmittedNote> submittedNoteList = submittedNoteRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
         List<Note> noteList = new ArrayList<>();
@@ -124,12 +134,31 @@ public class MeetingService {
             meetingMemberInfoDTOList.add(MeetingMemberInfoDTO.from(note.getMember()));
         }
 
-        return MeetingResponseDTO.of(meeting, meetingScript, meetingSummary, meetingKeywordDTOList, meetingMemberInfoDTOList);
+        return MeetingResponseDTO.of(meeting, meetingScript, meetingSummary, meetingKeywordDTOList, meetingMemberInfoDTOList, submittedNoteSummary);
     }
 
     @Transactional
-    public MeetingSummaryResponseDTO createSummary(Long studyId, Long meetingId) {
+    public synchronized MeetingSummaryResponseDTO createSummary(Long studyId, Long meetingId) {
         String summaryUrl = fastUrl + "/studies/summary";
+
+        Meeting meeting = meetingRepository.findByIdAndDeletedAtIsNull(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
+
+        if (meetingSummaryRepository.existsByMeetingAndDeletedAtIsNull(meeting)) {//기존의 미팅 sumary 데이터가 있을 경우
+            MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
+                    .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+
+            if (meetingSummary.getStatus() == IsolationEnum.USING) {//이미 누군가 meetingSummary값을 수정 중이라면
+                throw new BusinessException(ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS, ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS.getMessage());
+            }
+
+            //기존의 미팅요약을 삭제
+            meetingSummaryRepository.delete(meetingSummary);
+            meetingSummaryRepository.flush();
+        }
+
+        //새로운 미팅 생성(사용중인값으로 생성)
+        MeetingSummary meetingSummary = meetingSummaryRepository.saveAndFlush(MeetingSummary.of(meeting));
 
         String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findById(meetingId)
                         .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage())))
@@ -151,13 +180,10 @@ public class MeetingService {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
         HttpEntity<Object> requestEntity = new HttpEntity<>(jsonText, headers);
-        MeetingSummaryDTO meetingSummaryDTO = restTemplate.postForObject(summaryUrl, requestEntity, MeetingSummaryDTO.class);
 
         // db 저장
-        Meeting meeting = meetingRepository.findById(meetingId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
-        MeetingSummary meetingSummary = MeetingSummary.of(meeting, meetingSummaryDTO.getSummaryText());
-        meetingSummaryRepository.save(meetingSummary);
+        MeetingSummaryDTO meetingSummaryDTO = restTemplate.postForObject(summaryUrl, requestEntity, MeetingSummaryDTO.class);
+        meetingSummary.updateMeetingSummary(meetingSummaryDTO.getSummaryText());
         MeetingSummaryResponseDTO meetingSummaryResponseDTO = MeetingSummaryResponseDTO.from(meetingSummaryDTO);
 
         String message = "'" + meeting.getStudy().getTitle() + "' '" + meeting.getTopic() + "' 요약이 완료되었습니다.";
@@ -170,15 +196,20 @@ public class MeetingService {
     }
 
     public MeetingSummaryResponseDTO getMeetingSummary(Long meetingId) {
-        MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId).orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        if(meetingSummary.getStatus()==IsolationEnum.USING){
+            throw new BusinessException(ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS, ErrorCode.MEETING_SUMMARY_CAN_NOT_ACCESS.getMessage());
+        }
+
         return MeetingSummaryResponseDTO.getMeetingSummary(meetingSummary);
     }
 
     @Transactional
-    public MeetingSummaryResponseDTO updateCreateSummary(Long studyId, Long meetingId) {
+    public synchronized MeetingSummaryResponseDTO updateCreateSummary(Long studyId, Long meetingId) {
         String summaryUrl = fastUrl + "/studies/summary";
 
-        String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findById(meetingId)
+        String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findByIdAndDeletedAtIsNull(meetingId)
                         .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage())))
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()))
                 .getScriptContent();
@@ -205,7 +236,7 @@ public class MeetingService {
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
         meetingSummary.updateMeetingSummary(meetingSummaryDTO.getSummaryText());
 
-        Meeting meeting = meetingRepository.findById(meetingId)
+        Meeting meeting = meetingRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
 
         String message = "'" + meeting.getStudy().getTitle() + "' '" + meeting.getTopic() + "' 요약이 완료되었습니다.";
@@ -218,9 +249,12 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingSummaryResponseDTO updateMeetingSummary(MeetingSummaryUpdateDTO meetingSummaryUpdateDTO, Long meetingId) {
+    public synchronized MeetingSummaryResponseDTO updateMeetingSummary(MeetingSummaryUpdateDTO meetingSummaryUpdateDTO, Long meetingId) {
         MeetingSummary meetingSummary = meetingSummaryRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SUMMARY_NOT_FOUND, ErrorCode.MEETING_SUMMARY_NOT_FOUND.getMessage()));
+        meetingSummary.setUpdateStatus(IsolationEnum.USING);
+        meetingSummaryRepository.flush();
+
         meetingSummary.updateMeetingSummary(meetingSummaryUpdateDTO.getSummaryText());
 
         return MeetingSummaryResponseDTO.getMeetingSummary(meetingSummary);
@@ -230,20 +264,24 @@ public class MeetingService {
     public MeetingKeywordListDTO createMeetingKeyword(Long studyId, Long meetingId) {
         Meeting meeting = meetingRepository.findById(meetingId).orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         //이미 기존의 키워드가 존재 한다면?
-        if (meetingKeywordRepository.existsByMeetingAndDeletedAtIsNull(meeting)) {
-            List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
-
-            for (MeetingKeyword meetingKeyword : meetingKeywordList) {
-                meetingKeywordRepository.delete(meetingKeyword);
-            }
-        }
-
-        String keywordUrl = fastUrl + "/studies/keyword";
 
         String originalText = meetingScriptRepository.findByMeetingAndDeletedAtIsNull(meetingRepository.findById(meetingId)
                         .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage())))
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_SCRIPT_NOT_FOUND, ErrorCode.MEETING_SCRIPT_NOT_FOUND.getMessage()))
                 .getScriptContent();
+
+        if (meetingKeywordRepository.existsByMeetingAndDeletedAtIsNull(meeting)) {
+            List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
+
+            for (MeetingKeyword meetingKeyword : meetingKeywordList) {
+                if(meetingKeyword.getStatus()==IsolationEnum.USING)
+                    throw new BusinessException(ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS, ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS.getMessage());
+                meetingKeywordRepository.delete(meetingKeyword);
+            }
+            meetingKeywordRepository.flush();
+        }
+
+        String keywordUrl = fastUrl + "/studies/keyword";
 
         //originalText = originalText.replace(".", ".\n");
 
@@ -268,7 +306,7 @@ public class MeetingService {
         // db 저장
 
         for (int i = 0; i < MeetingKeywordResponseDTO.getKeyword().size(); i++) {
-            meetingKeywordRepository.save(MeetingKeyword.of(meeting, MeetingKeywordResponseDTO.getKeyword().get(i)));
+            meetingKeywordRepository.saveAndFlush(MeetingKeyword.of(meeting, MeetingKeywordResponseDTO.getKeyword().get(i)));
         }
 
         List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
@@ -290,12 +328,15 @@ public class MeetingService {
     }
 
     public MeetingKeywordListDTO getMeetingKeyword(Long meetingId) {
-        Meeting meeting = meetingRepository.findById(meetingId)
+        Meeting meeting = meetingRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
         List<MeetingKeywordDTO> meetingKeywordDTOList = new ArrayList<>();
 
         for (MeetingKeyword meetingKeyword : meetingKeywordList) {
+            if(meetingKeyword.getStatus()==IsolationEnum.USING){
+                throw new BusinessException(ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS, ErrorCode.MEETING_KEYWORD_CAN_NOT_ACCESS.getMessage());
+            }
             MeetingKeywordDTO meetingKeywordDTO = MeetingKeywordDTO.from(meetingKeyword);
             meetingKeywordDTOList.add(meetingKeywordDTO);
         }
@@ -303,7 +344,7 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingKeywordListDTO updateMeetingKeyword(Long meetingId, UpdateMeetingKeywordListDTO updateMeetingKeywordListDTO) {
+    public synchronized MeetingKeywordListDTO updateMeetingKeyword(Long meetingId, UpdateMeetingKeywordListDTO updateMeetingKeywordListDTO) {
         for (UpdateMeetingKeywordDTO updateMeetingKeywordDTO : updateMeetingKeywordListDTO.getUpdateMeetingKeywordList()) {
             MeetingKeyword meetingKeyword = meetingKeywordRepository.findByIdAndDeletedAtIsNull(updateMeetingKeywordDTO.getGroupKeywordId())
                     .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_KEYWORD_NOT_FOUND, ErrorCode.MEETING_KEYWORD_NOT_FOUND.getMessage()));
@@ -311,7 +352,7 @@ public class MeetingService {
             meetingKeyword.updateMeetingKeyword(updateMeetingKeywordDTO.getKeyword());
         }
 
-        Meeting meeting = meetingRepository.findById(meetingId)
+        Meeting meeting = meetingRepository.findByIdAndDeletedAtIsNull(meetingId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND, ErrorCode.MEETING_NOT_FOUND.getMessage()));
         List<MeetingKeyword> meetingKeywordList = meetingKeywordRepository.findAllByMeetingAndDeletedAtIsNull(meeting);
         List<MeetingKeywordDTO> meetingKeywordDTOList = new ArrayList<>();
